+++ excalidraw-app/data/StorageBackend.ts
@@ -1,9 +1,7 @@
-import type { SyncableExcalidrawElement } from ".";
-import type {
-  ExcalidrawElement,
-  FileId,
-} from "@excalidraw/element/types";
+import type { ExcalidrawElement, FileId } from "@excalidraw/element/types";
 import type { AppState, BinaryFileData } from "@excalidraw/excalidraw/types";
+
+import type { SyncableExcalidrawElement } from ".";
 import type Portal from "../collab/Portal";
 import type { Socket } from "socket.io-client";
 
@@ -56,4 +54,3 @@ export interface StoredScene {
   iv: Uint8Array;
   ciphertext: ArrayBuffer;
 }
-